Add stub to Jotz.js and refactor BlockMenu.js
