+++ src/renderer/components/content/editor/aceConfig.js
@@ -16,7 +16,7 @@ var Ace = require('brace');
 //require('brace/mode/cobol');
 //require('brace/mode/coffee');
 //require('brace/mode/coldfusion');
-require('brace/mode/csharp');
+//require('brace/mode/csharp');
 require('brace/mode/css');
 //require('brace/mode/curly');
 //require('brace/mode/d');
@@ -40,7 +40,7 @@ require('brace/mode/css');
 //require('brace/mode/handlebars');
 //require('brace/mode/haskell');
 //require('brace/mode/haxe');
-//require('brace/mode/html');
+require('brace/mode/html');
 //require('brace/mode/html_ruby');
 //require('brace/mode/ini');
 //require('brace/mode/io');
@@ -49,8 +49,8 @@ require('brace/mode/css');
 //require('brace/mode/java');
 require('brace/mode/javascript');
 require('brace/mode/json');
-require('brace/mode/jsoniq');
-require('brace/mode/jsp');
+//require('brace/mode/jsoniq');
+//require('brace/mode/jsp');
 require('brace/mode/jsx');
 //require('brace/mode/julia');
 //require('brace/mode/latex');
@@ -64,7 +64,7 @@ require('brace/mode/jsx');
 //require('brace/mode/luapage');
 //require('brace/mode/lucene');
 //require('brace/mode/makefile');
-//require('brace/mode/markdown');
+require('brace/mode/markdown');
 //require('brace/mode/matlab');
 //require('brace/mode/mel');
 //require('brace/mode/mushcode');
@@ -76,7 +76,7 @@ require('brace/mode/jsx');
 //require('brace/mode/perl');
 //require('brace/mode/pgsql');
 //require('brace/mode/php');
-require('brace/mode/plain_text');
+//require('brace/mode/plain_text');
 //require('brace/mode/powershell');
 //require('brace/mode/praat');
 //require('brace/mode/prolog');
@@ -88,7 +88,7 @@ require('brace/mode/plain_text');
 //require('brace/mode/rhtml');
 //require('brace/mode/ruby');
 //require('brace/mode/rust');
-//require('brace/mode/sass');
+require('brace/mode/sass');
 //require('brace/mode/scad');
 //require('brace/mode/scala');
 //require('brace/mode/scheme');
@@ -104,7 +104,7 @@ require('brace/mode/plain_text');
 //require('brace/mode/svg');
 //require('brace/mode/tcl');
 //require('brace/mode/tex');
-//require('brace/mode/text');
+require('brace/mode/text');
 //require('brace/mode/textile');
 //require('brace/mode/toml');
 //require('brace/mode/twig');
